github: fix some failing workflows
Snapcraft 7.x doesn't implement the new CRAFT variables which are
available in snapcraft 8.x, so we'll switch them back for some jobs.

We can reenable the Polarfire build as this was the cause of the failure!

We'll also disable two builds for now, as the runner runs out of
space during the build. We'll have to figure out if we can get more
disk space, or (more likely) dispatch a build to Launchpad when we're
building the kernel from source instead of staging a deb.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>